Instead of Promises, use old callbacks
Really, you shouldn't really be using these callbacks
anyway because your logic should not rely on an image asset
being loaded - so this should really be more of a fire and
forget.

But just incase, here you go.

We move away from Promises because an uncaught error will show
up in the console tab and be generally annoying
even though its not 'really' an error